Added an typing Practice Website
There was two typing practice site included in TOP. But I thought they are not that much helpful as (https://www.typing.com/) So I added this website as an alternative typing practice site. I used it personally and believe it can help students to get more boost in their typing.