Flexbox Alignment: Update resources
This comprehensive MDN article provides a solid foundation for understanding the basics of CSS Flex-box. It extensively covers the concept of axes while also briefly introducing alignment. To optimize the learning flow, I have moved this section to the previous lesson on flex-box axes, where students will receive a brief introduction to alignment. Additionally, in order to avoid duplication, I have removed the Kevin Powell video from this lesson since it is already mentioned in the Growing and Shrinking lesson. Instead, I have included the cheat sheet from the previous lesson in this current lesson.