Flexbox Axes: Update resource
To enhance the learning experience, it is advantageous to switch the placement of resources. The old resource, a cheat sheet encompassing all Flex-box properties, is better suited for the final lesson before the project. Meanwhile, the new resource, originally intended for the next lesson, has been moved here, allowing for a more detailed examination of the concepts discussed in the current lesson on "flex-box axes".