Update deployment.md to add info on env. variables
Added a bullet after step two of the assignment as this caused me hours of strife.  The Fly logs indicated the environment variable wasn't being read, but after scouring the dotenv docs and the tutorial guide for launching an app on Fly.io, there was no mention of needing special handling for environment variables anywhere so I kept thinking it just wasn't being instantiated early enough.  Additionally, when I finally found a stackoverflow answer that mentioned the Fly docs for handling secrets, nowhere do the docs mention the `import` command; they only describe how to manually set a secret in the command line (rather than importing from a .env file), which doesn't align with the fact that the lesson specifically suggests using Dotenv, hence my decision to use that as the example.