Update testing_basics.md
I noticed that the additional resources section for the testing basics in javascript course is currently empty. I'd like to contribute by suggesting a helpful resource for further learning. It is a 13 minute youtube video by Web Dev Simplied showing how to get started with testing in JavaScript using Jest.