Update links-and-images.md
Added a little bit of information that when measurement units for the "height" and "width" attributes are not explicitly defined the values are assumed to be in pixels (px) by default in HTML.
There isn't anything mentioned about this in the lesson. So it could have been confusing for beginners.