Update instructions to include setting up Jest
The lesson for testing node/express applications uses Jest but does not make clear that Jest is needed to make the test work  in this particular example. I included some instructions to install and setup Jest and to take a look back at the previous lessons if one does not remember how to.