Remove redundant <> tags, format <svg> tags
In the 3rd bullet point - camelCase Most things - code samples provide redundant <> tags </>. It can be misleading for learners, whether should they put the tags everywhere or not. Therefore, I suggest it be removed.

The other change is, some of the <svg > tags have blank space just after the 'svg' and before '>' - it also can be misleading, and seems to be a little typo.