Update fundamentals-1.md
I have added a W3Schools link to the JavaScript tutorial which explains the use of 'const', a detail not previously included. This article clarifies that 'const' is used to declare arrays and objects. While the type of the variable won't change, the data within it can be modified if it's an array or an object.