JavaScript/webpack: Add important note about WebPack fonts (#27300)
* add important note about WebPack fonts

I posted an answer in the discord for this issue and have had numerous messages over months thanking me for pointing them in the right direction, so I believe it is worth adding. It is a common pitfall that isn't pointed out in the WebPack tutorial.

* amend an important note about fonts

Important note addressing common pitfall of using ttf fonts in webpack. Changed from original PR to link directly to SO, and to briefly describe the linked webpage now.

* Update javascript/organizing_your_javascript_code/webpack.md

Co-authored-by: Kevin <kevinmulhern23@gmail.com>

* Update webpack.md

Co-authored-by: Kevin <kevinmulhern23@gmail.com>

* Update webpack.md

Co-authored-by: Kevin <kevinmulhern23@gmail.com>

* Update webpack.md

Co-authored-by: Kevin <kevinmulhern23@gmail.com>

* Update webpack.md

Co-authored-by: Kevin <kevinmulhern23@gmail.com>

* Update webpack.md

Co-authored-by: Kevin <kevinmulhern23@gmail.com>

* Update webpack.md

Co-authored-by: Kevin <kevinmulhern23@gmail.com>

---------

Co-authored-by: Kevin <kevinmulhern23@gmail.com>