context lesson: Fix linting errors
Purposefully left code block delimiters as tildes - this is something
that is part of an open issue to fix these.
The context lesson is part of a group of 3 lessons which have not yet
been assigned (as of this commit).
All other lint errors should be addressed in this commit.