Dynamic User Interface Interactions: Clarify Image Slider instructions
The CSS transitions video is important to step 3 of the instructions for the Image Slider exercise.

It was not implied as imperative to the Image Slider exercise to watch the video on transitions,
but it is necessary to step 3 as an introduction to CSS transitions.

Changes:
Line 9: Clarify that the video demonstrates CSS transitions.
Line 42: Reference the transitions video for added clarity.