Make transitions video crucial to the Image Slider
This is a simple issue that has not yet been opened.

It was not implied as imperative to the Image Slider exercise to watch the video on transitions,
but it is necessary to Image Slider step 3 as an introduction to CSS transitions.

Changes:
Line 9: Clarify that the video demonstrates CSS transitions.
Line 42: Reference the transitions video for added clarity.