Improve readability in rule code
Opted to return token.map as is on line 19 instead of the same but with 1
subtracted from token.map[1].
The only difference this makes is include the closing backticks line in
the isWithinIgnoredFence range check, which will have no effect on the
outcome of the check.
The benefit is not needing to explain why we need to subtract 1 from
token.map[1] only. This allows everything to remain intuitive, and the
only deviation being the onError.lineNumber value which has a comment to
explain its 1-index requirement.

Regexr test links added.

Co-authored-by: Eric Olkowski <thatblindgeye@gmail.com>