Helpful vids for visualization of Event Loop and Promises
I found these videos extra helpful in further visualizing how JavaScript works in the background with the Browser. 

I included it in the Assignment section because I felt it was that useful, but can decide if want to include them or not. Also put them in position 4, 5 because I felt that learning order was best.

Thank you for all that y'all do!