Account for surrounding lines being spaces only
Without accounting for this, the fixer will insert a blank line and
MD009 can't fix the trailing whitespace, leaving us with a double line
break *and* trailing whitespace still present.

Accounting for this allows MD009 to fix trailing white space, and the
result to be fully fixed without extra line breaks.