Update Node crash course video
The current link for the Node crash course video pointed to a version that is 5 years old. I have updated the link to the most recent version of the same course, which was updated last month. This ensures that users are learning from the most up-to-date material available.

Details of the change:
- Old video link: [https://www.youtube.com/watch?v=fBNz5xF-Kx4]
- New video link: [https://www.youtube.com/watch?v=32M1al-Y6Ag&t]

This update provides a more current and relevant educational resource, improving the overall quality of the curriculum.