Edit working-with-text.md, small change to the wording of a sentence under 'Nesting and indentation'.
This is a small edit to the wording of a sentence in the 'Nesting and indentation' section. I changed: "It is recommended to indent any child elements by two spaces." to: "In our examples, we have indented any child elements by two spaces."

Here is a link to a Discord conversation that should provide some context and reasoning for the change:  https://discord.com/channels/505093832157691914/690589233919819797/1255563153329098813 

In short, this changed version is more fitting since the amount of spaces per indentation is largely a matter of personal preference.