Fix typo in Controllers Lesson
Corrected "fnuctions" to "functions"

Original text:
"...So you need to make sure and be aware that your middleware fnuctions
are placed in the correct order. As an example, some packages
have middleware functions that changes the Request object,
and as a result, these middleware functions should be placed
at the very top of your application in order for you to be able
to see their changes in all of your middleware functions below it...."

Fixed text:
"....So you need to make sure and be aware that your middleware
functions are placed in the correct order. As an example,
some packages have middleware functions that changes the Request object,
and as a result, these middleware functions should be placed
at the very top of your application in order for you to be
able to see their changes in all of your middleware functions below it...."