Update Views: Make writing style more imperative
This PR addresses learners' frustrations regarding the difficulty
understanding what text is intended as hypothetical examples and what
text is intended as instructions to follow, due to the change in mood
from passive to imperative. This PR has changed the passive style text
to imperative where relevant, as well as fleshed out the `user.forEach`
code example to indicate what the learner should do more clearly.