Links and Images: Add additional resource link for better understanding image formats  (#28612)
* Add additional resource link for understanding image formats

Added a link to an imagekit.io article in the "Additional Resources" section of the "Links and Images" lesson. The article compares JPEG, PNG, GIF, and WebP formats, helping learners choose the right format for web use.

* Improve link text for image formats resource

Updated the link text in the "Additional Resources" section of the "Links and Images" lesson to be more descriptive. The link now clearly indicates that it leads to an article on the best image formats for web use, ensuring better accessibility and compliance with guidelines.

* Remove "this" word from a link description

Ensuring compliance with guidelines.