Installing Ruby: Update installing_ruby.md (#28624)
* Update installing_ruby.md

Version 1.3.0 of rbenv has a feature change where rbenv init comes from the bin directory instead of right in the ~/.rbenv directory and now modifies shell initialization files instead of only printing instructions. 

The lesson needs to now either specify whether to use an older version, or if using the latest version, the terminal must be refreshed after this step. 

I am fresh out of Foundations and unsure how this change may affect any subsequent lessons in the course.

* Update installing_ruby.md

A blank line was needed under the code block. Simplified instructions on closing and opening the terminal.