Clarifying instructions on step 4.1 for VMs
As I was making the course, the instruction "... click Settings -> Display" wasn't clear enough as to where to find it. 

As it follows the "...power off VM" segment, I thought I had to go to my Windows settings, but it is related to VirtualBox settings.

I made this clearer as to where to find this "Settings -> Display" by adding "in VirtualBox".