Add a link about methods return values
When I finished the ruby_basics exercise, I looked at the solutions and noticed that the functions (or methods) didn’t have a return statement. I found this odd, which led me to the article I’ve linked, which explains that Ruby methods don’t require an explicit return statement.