Ruby: Project: Caesar Cipher: Fix Ruby Caesar Cipher code example (#29107)
* Caesar_cipher code example showed right shift, not left.

A left shift of 5 for the string "What a string" should yield the value "Rcvo v nomdib!". The output in the example shows a right shift of 5.

At the very least, the example from Wikipedia shows a left shift of three, "D->A" and "E->B" while the code example shows a right shift. The examples should be consistent on the page.

* Add instructions for right shift

* Add to project instructions using a right shift
* Add to quick tips that Wikipedia shows a left shift
* Fix lower case ceaser