Update transforms.md
The description of order of transformations was incorrect. MDN says that composition is right to left, which is talking about the inner function g of f(g(x)) being done first. This is equivalent to transform: g() f(), so when laid out the transform actually goes from left to right when reading it. I have changed the description to describe this change too.