Rewrite assignment to offer extra credit
I propose this change in the task statement to offer an additional voluntary exercise, which would be to extend the game by adding the Lizard and Spock elements.

The reasons for making this proposal are three:
1. To make my little contribution to TOP.
2. To show that there is a contradiction between what is said in the lesson statement:

"When making interactive projects, like this one, you might be tempted to add more features, improve interactivity, user experience, design and styling of your website, and so on.
We recommend not doing this and reserving this effort for your portfolio projects."

...with other assignments along TOP; for example, in the Ruby course there are some extra credits, or voluntary exercises.
3. To complete the exercise "Using Git in the Real World", from the Ruby course.

If you don't accept the pull request, at least I would like you to tell me if I have done this exercise properly (that is to say, if I have followed the git flow correctly).

If you want, you can visit my Rock, Paper, Scissors, Lizard Spock project in TOP:

https://github.com/TheMiki666/rock_paper_scissors

Thanks for your time.