Update project_basic_informational_site.md
Under the instructions of the Project: Basic Informational Site assignment, on instruction two (2), the bulleted-list hints, I propose that the file extension should be added to the hints to provide a clearer meaning to readers. With the way it's right now, readers are being instructed to access the file pathname which does not exist. With my server running I tried to access the links or pages as described in the assignment instructions, it's the 404.html page that is always returned. The pages are served correctly when I add the `.html` extension to the file name. Rather than this:

- [localhost:8080](http://localhost:8080/) should take users to index.html
- [localhost:8080/about](http://localhost:8080/about) should take users to about.html
- [localhost:8080/contact-me](http://localhost:8080/contact-me) should take users to contact-me.html
- 404.html should display any time the user tries to go to a page not listed above.

It should be this way:

- [localhost:8080](http://localhost:8080/) should take users to index.html
- [localhost:8080/about.html](http://localhost:8080/about.html) should take users to about.html
- [localhost:8080/contact-me.html](http://localhost:8080/contact-me.html) should take users to contact-me.html
- 404.html should display any time the user tries to go to a page not listed above.