Add detail about event callback parameter to DOM manipulation section
This section was confusing to me when I read it the first time. It wasn't clear to me whether `e` was a magical keyword or an arbitrary variable name. In theory, I have enough prior exposure to callbacks and event handlers that I could have worked it out on my own, so I imagine it might be even more confusing for learners without that experience.

Some of the other readings in the assignment section also build on this, but the MDN "Introduction to events" isn't among them, and I felt that having it here would be helpful for learners who stumbled where I did.