easy-fix: Incorrect and awkward usage of transitive verb 'refresh' (#29396)
Convention: 'To refresh yourself' typically implies a physical action,
such as 'to refresh yourself with a cold shower'. There are several
more commonly used colloquialisms for 'review', listed below:

  1. Refresh your memory
  2. Refresh your knowledge
  3. Give yourself a refresher

Grammar: Since the object being acted on is the subject's 'memory',
it follows that 'memory' should be the object being refreshed. In
this case however, 'refresh yourself' incorrectly moves the object
pointer from 'memory' to 'you'.