refactor: Roll Linux and ChromeOS instructions into one
ChromOS now comes with Linux Development Environment which includes APT.
Installation can be done using APT as on a regular Linux system.
Instructions for Linux can now be used by ChromeOS users as well.
Guide on turning on Linux in ChromeOS is linked as reference.
Article on installation using source is linked as legacy approach for older Chromebooks.