Databases and SQL: Fix grammar and clarify SQL concepts (#29611)
* Fix minor grammar issues

* Fix incorrect reference to 'whole database'

The sentence, "If you just do an action on a table without specifying conditions, it will apply to the whole database and you’ll probably break something" mentions that the action is on a table. Hence, it would affect the table, not the whole database.

* Clarify INSERT syntax with and without columns

* Clarify when * can be used in aggregate functions

* Correct the reference to HAVING

It's incorrectly referred to as a function, but `HAVING` is a clause.

* Surround fenced code blocks with blank lines

* Undo lowercasing of E.g. to e.g.

* Change reference to HAVING in knowledge check section

HAVING is a clause, not a function.

* Unuse semicolon and make two sentences

* Use inline code formatting for table and column names

* Clarify INSERT INTO usage

Revised explanation to mention that omitting column names is possible but discouraged, without showing an example.

* Update the id of HAVING function section

The reference to `HAVING` was recently updated from "function" to "clause" and the id needed updating as well as the knowledge check link.