Class Based Components: Improve code and grammar (#29217)
* Rename "todo" to "prevTodos" for clarity in state update

* Improve input-label connection

The label element has a `htmlFor` attribute set to "task-entry", but the input element doesn't have an `id` attribute. I have given the input an id that matches the value in the `htmlFor` attribute of the label.

* Fix grammar in comments and explanation

* Revise Count component description in assignment section

* Change To-Dos to Todos in comments

This is to stay consistent with the naming used in the code and text in the lesson.

Co-authored-by: Kevin <kevinmulhern23@gmail.com>

---------

Co-authored-by: Kevin <kevinmulhern23@gmail.com>