React Router lesson: Fix link to useOutletContext documentation
The currently live link (https://api.reactrouter.com/v7/functions/react_router.useOutletContext.html) goes to the reference documentation page, which doesn't include or link to helpful information for learners. Unfortunately, the latest (7.3.6) and dev documentation versions *also* have bare-bones documention that only gives the hook signature (e.g. https://reactrouter.com/api/hooks/useOutletContext). However, the 6.30.1 docs provide useful examples, and I suspect this was the documentation the lesson originally intended to link to.

I suspect that as the React Router docs get updated, this link will need to change again, but for now, if the lesson contains a link to official React Router documentation for this hook, this should be the page it links to.