Working with APIs: Remove manual cors mode option (#28733)
`mode: 'cors'` is already the default in browsers.
The CORS section also doesn't include anything too vital, given the detour to explore CORS is treated as optional, and would be fairly out of scope of this lesson anyway.
The section's original intent was just to make sure things work, but they already do without the manual option.