Update advanced_selectors.md to remove redundant expression
"specific" and "finely grained" mean the same thing so it's probably enough to have one. "specific" is shorter, so I chose that version. 

Also, if keeping "finely grained" is desired, it might be better to change it to the idiomatic form which is "fine-grained way" rather than "finely grained way".